+++ resources/types.go
@@ -5,6 +5,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
+	networkingv1 "k8s.io/api/networking/v1"
 	policyv1 "k8s.io/api/policy/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -24,6 +25,7 @@ type DeploymentValues struct {
 	InitContainers      []Container
 	Ingress             *schema.Ingress
 	HTTPRoute           *schema.HTTPRoute
+	NetworkPolicies     map[string]networkingv1.NetworkPolicySpec
 	Volumes             map[string]schema.Volume
 	Sidecars            map[string]schema.Container
 	PreDeploymentJob    *PreDeploymentJob