+++ CHANGELOG.md
@@ -14,6 +14,31 @@ For clarity we use following emoji for the changes:
 
 ## [unreleased]
 
+## [1.4.0] - 2026-03-13
+### :star: Added
+- support for creating `NetworkPolicies`
+  - these behave exactly like `configMaps`
+  - e.g.
+    ```yaml
+    networkPolicies:
+      deny-all:
+        podSelector:
+          matchLabels:
+            app: foo
+        policyTypes: [Ingress, Egress]
+      allow-some:
+        podSelector:
+          matchLabels:
+            app: foo
+        policyTypes: [Ingress]
+        ingress:
+          - from:
+            - namespaceSelector:
+                matchLabels:
+                  kubernetes.io/metadata.name: ingress-nginx
+    ```
+  - this will render 2 `NetworkPolicies` - `{service}--{component}--{env}-deny-all` and `{service}--{component}--{env}-allow-some` in the given namespace, and the value of each is the `NetworkPolicySpec`
+
 ## [1.3.0] - 2026-02-06
 
 ### :star: Added
@@ -70,7 +95,8 @@ For clarity we use following emoji for the changes:
 
 ### :star: Moved the project to public GitHub repository! :rocket:
 
-[unreleased]: https://github.com/ProRocketeers/yoke-chart/compare/1.3.0...HEAD
+[unreleased]: https://github.com/ProRocketeers/yoke-chart/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/ProRocketeers/yoke-chart/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/ProRocketeers/yoke-chart/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/ProRocketeers/yoke-chart/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/ProRocketeers/yoke-chart/compare/1.0.0...1.1.0